# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Agent Security Platform API
  version: 1.0.0
  description: Stable product API contract.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
paths:
  /api/v1/session/start:
    get:
      operationId: startSession
      summary: Start a repository-owned browser identity flow
      security: []
      parameters:
        - name: return_to
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 2048, pattern: '^/(?!/)' }
      responses:
        '302': { description: Redirect to the configured identity authorization endpoint. }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
...
  /api/v1/session/callback:
    post:
      operationId: completeSessionCallback
      summary: Exchange an identity-provider callback for a host-only browser session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCallbackInput' }
      responses:
        '200':
          description: Session established with a Secure, HttpOnly, SameSite=Lax __Host-zasp_session cookie and the state-bound return path.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Host-only __Host-zasp_session cookie; Secure, HttpOnly, Path=/, SameSite=Lax.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionCallbackResult' }
        '400': { $ref: '#/components/responses/ProductErrorResponse' }
        default: { $ref: '#/components/responses/ProductErrorResponse' }
...
security:
  - BrowserSession: []
    BrowserExpectedScope: []
  - ProductAPIToken: []
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-zasp_session
      description: Secure, HttpOnly, SameSite=Lax host-only human browser session.
    BrowserExpectedScope:
      type: apiKey
      in: header
      name: X-Zasp-Expected-Scope
      description: Exact canonical organization/workspace/environment scope captured by the browser before a scoped request. The server compares it with the authenticated session scope before data access; ProductAPIToken scope remains token-owned.
    ProductAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ProductAPIToken
      description: Product automation token.
...
  schemas:
...
    SessionCallbackInput:
      type: object
      additionalProperties: false
      required: [provider_token, state]
      properties:
        provider_token: { type: string, minLength: 1, maxLength: 4096 }
        state: { type: string, minLength: 32, maxLength: 512 }
    SessionCallbackResult:
      type: object
      additionalProperties: false
      required: [return_to]
      properties:
        return_to: { type: string, minLength: 1, maxLength: 2048, pattern: '^/(?!/)' }
...
    ProductID:
      type: string
      minLength: 40
      maxLength: 40
      pattern: '^pid_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      description: Canonical product identifier.
...
    ProductError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - correlation_id
        - retryable
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z][a-z0-9]*(?:_[a-z0-9]+)*$'
        message:
          type: string
          minLength: 1
          maxLength: 512
          pattern: '^(?!\s)(?!.*\s$)[^\x00-\x1F\x7F-\x9F]+$'
        correlation_id:
          $ref: '#/components/schemas/ProductID'
        retryable:
          type: boolean
  responses:
    ProductErrorResponse:
      description: Request rejected with a product error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'